Fix Shortcut Manager target-column width type

+++ src/ui/ShortcutManagerWindow.cpp
@@ -886,7 +886,8 @@ UpdateColumnWidths() {
     const int target =
         std::max(
             minimumTarget,
-            client.right -
+            static_cast<int>(
+                client.right) -
                 keywords -
                 name -
                 type -
